Based Letradical Additions
So essentially, I've added / rewrote two things.

The first is just how the output is formatted, I replaced the constant of MAX_STR_LEN with a function that reads the longest path name and assigns that to the value instead.

Secondly I added the ability to ignore directories with `-i` / `-ignore`:

micelia -i test -i another_test /path/to/directory